+++ src/components/scrollable_layout.ts
@@ -4,21 +4,23 @@ const defaultScrollbar = {
   "scrollbarWidth": "initial",
   "scrollbarColor": "initial",
   "&::-webkit-scrollbar": { all: "initial" },
-  "&::-webkit-scrollbar-thumb": { all: "initial", background: "gray" },
+  "&::-webkit-scrollbar-thumb": {
+    all: "initial",
+    background: "#00000088",
+  },
   "&::-webkit-scrollbar-track": { all: "initial" },
 };
 
 export const Container = styled("div", {
   position: "relative",
   height: "100%",
-  ...defaultScrollbar,
+  userSelect: "none",
 });
 
 export const ScrollableLayout = styled("div", {
   // chrome user-agent style override
   outline: 0,
   position: "relative",
-  backgroundColor: "#eee",
   width: "100%",
   height: "100%",
   display: "flex",
@@ -35,6 +37,14 @@ export const ScrollableLayout = styled("div", {
         top: 0,
         bottom: 0,
         overflow: "auto",
+      },
+    },
+    dark: {
+      true: {
+        "&::-webkit-scrollbar-thumb": {
+          all: "initial",
+          background: "#ffffff88",
+        },
       },
     },
   },